Added individual chat fetching
