+++ src/app/api/chat/route.ts
@@ -1,11 +1,7 @@
 import { auth } from '@/app/auth';
+import parseDbMessages from '@/lib/chat/parseDbMessages';
 import client from '@/lib/db';
-import {
-  Chat,
-  ChatMessage,
-  ChatPostBody,
-  MongoDBChatMessage,
-} from '@/lib/types/chat';
+import { ChatPostBody, MongoDBChatMessage } from '@/lib/types/chat';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET() {
@@ -19,7 +15,7 @@ export async function GET() {
   const chatsCollection = client.db('huddle-chat').collection('chats'); // Get the chats collection
 
   // Get all chat messages for the user and join with the user collection
-  const chatMessages = await chatsCollection
+  const chatMessages = (await chatsCollection
     .aggregate([
       {
         $match: { userId },
@@ -41,65 +37,9 @@ export async function GET() {
         $unwind: { path: '$user', preserveNullAndEmptyArrays: true },
       },
     ])
-    .toArray();
+    .toArray()) as MongoDBChatMessage[];
 
-  // Format the chat messages to match the ChatMessage type
-  const formattedMessages = chatMessages
-    .map((message) => ({
-      chatId: message.chatId,
-      chatType: message.chatType,
-      messageId: message._id.toString(),
-      message: message.message || undefined,
-      reaction: message.reaction || undefined,
-      replyId: message.replyId || undefined,
-      timestamp: message.timestamp,
-      user: {
-        id: message.user?._id.toString(),
-        name: message.user?.name,
-        image: message.user?.image,
-      },
-    }))
-    .sort((a, b) => a.timestamp - b.timestamp) as ChatMessage[];
-
-  // Split chat messages into their chat rooms (removing reported messages)
-  const chatRooms: Chat[] = Object.values(
-    formattedMessages.reduce(
-      (acc, message) => {
-        if (message.reported) return acc;
-        if (!acc[message.chatId])
-          acc[message.chatId] = {
-            chatId: message.chatId,
-            chatType: message.chatType,
-            messages: [],
-          };
-
-        if (!message.message && message.reaction) {
-          const referencedMessage = acc[message.chatId].messages.find(
-            (m) => m.messageId === message.replyId,
-          );
-          if (referencedMessage) {
-            if (!referencedMessage.reactions[message.reaction])
-              referencedMessage.reactions[message.reaction] = 0;
-            referencedMessage.reactions[message.reaction]++;
-          }
-          return acc;
-        }
-
-        acc[message.chatId].messages.push({
-          messageId: message.messageId,
-          message: message.message as string,
-          reactions: {},
-          replyId: message.replyId,
-          timestamp: message.timestamp,
-          user: message.user,
-        });
-        return acc;
-      },
-      {} as {
-        [chatId: string]: Chat;
-      },
-    ),
-  );
+  const chatRooms = parseDbMessages(chatMessages); // Parse the chat messages
 
   return NextResponse.json(chatRooms);
 }
